move old unused code to deepqmc.extra
