+++ setup.cfg
@@ -29,7 +29,7 @@ omit =
     */deepqmc/defaults.py
     */deepqmc/torchext/sloglindet.py
     */deepqmc/torchext/bdet.py
-    */deepqmc/utils.py
+    */deepqmc/torchext/cuda.py
     */deepqmc/extra/**.py
 
 [pydocstyle]
